Drawer navigation initial changes
